feat: add list part api to node and browser
