+++ lib/browser/bucket.js
@@ -56,6 +56,15 @@ proto.useBucket = function useBucket(name, region) {
   return this;
 };
 
+proto.setBucket = function useBucket(name) {
+  this.options.bucket = name;
+  return this;
+};
+
+proto.getBucket = function getBucket() {
+  return this.options.bucket;
+};
+
 proto.putBucket = function* putBucket(name, region, options) {
   var params = this._bucketRequestParams('PUT', name, '', options);
   if (region) {
