+++ lib/browser/client.js
@@ -109,10 +109,12 @@ merge(proto, require('./object'));
 //  * Bucket operations
 //  */
 // merge(proto, require('./bucket'));
+//multipart upload
+merge(proto, require('./managed_upload'));
 /**
  * Multipart operations
  */
-merge(proto, require('./multipart'));
+merge(proto, require('../common/multipart'));
 
 /**
  * Common module
