+++ lib/browser/managed_upload.js
@@ -1,10 +1,7 @@
 'use strict';
 
 // var debug = require('debug')('ali-oss:multipart');
-// var fs = require('fs');
 var is = require('is-type-of');
-// var destroy = require('destroy');
-// var eoe = require('end-or-error');
 var util = require('util');
 var path = require('path');
 var mime = require('mime');
@@ -166,153 +163,7 @@ proto._resumeMultipart = function* _resumeMultipart(checkpoint, options) {
   return yield this.completeMultipartUpload(name, uploadId, doneParts, options);
 };
 
-/**
- * List the on-going multipart uploads
- * https://help.aliyun.com/document_detail/31997.html
- * @param {Object} options
- * @return {Array} the multipart uploads
- */
-proto.listUploads = function* listUploads(query, options) {
-  options = options || {};
-  options.subres = 'uploads';
-  var params = this._objectRequestParams('GET', '', options)
-  params.query = query;
-  params.xmlResponse = true;
-  params.successStatuses = [200];
 
-  var result = yield this.request(params);
-  var uploads = result.data.Upload || [];
-  if (!Array.isArray(uploads)) {
-    uploads = [uploads];
-  }
-  uploads = uploads.map(function (up) {
-    return {
-      name: up.Key,
-      uploadId: up.UploadId,
-      initiated: up.Initiated
-    };
-  });
-
-  return {
-    res: result.res,
-    uploads: uploads,
-    bucket: result.data.Bucket,
-    nextKeyMarker: result.data.NextKeyMarker,
-    nextUploadIdMarker: result.data.NextUploadIdMarker,
-    isTruncated: result.data.IsTruncated === 'true'
-  };
-};
-
-/**
- * Abort a multipart upload transaction
- * @param {String} name the object name
- * @param {String} uploadId the upload id
- * @param {Object} options
- */
-proto.abortMultipartUpload = function* abortMultipartUpload(name, uploadId, options) {
-  this.cancel();
-  options = options || {};
-  options.subres = {uploadId: uploadId};
-  var params = this._objectRequestParams('DELETE', name, options);
-  params.successStatuses = [204];
-
-  var result = yield this.request(params);
-  return {
-    res: result.res
-  };
-};
-
-/**
- * Initiate a multipart upload transaction
- * @param {String} name the object name
- * @param {Object} options
- * @return {String} upload id
- */
-proto.initMultipartUpload = function* initMultipartUpload(name, options) {
-  options = options || {};
-  options.headers = options.headers || {};
-  this._convertMetaToHeaders(options.meta, options.headers);
-
-  options.subres = 'uploads';
-  var params = this._objectRequestParams('POST', name, options);
-  params.mime = options.mime;
-  params.xmlResponse = true;
-  params.successStatuses = [200];
-
-  var result = yield this.request(params);
-
-  return {
-    res: result.res,
-    bucket: result.data.Bucket,
-    name: result.data.Key,
-    uploadId: result.data.UploadId
-  };
-};
-
-/**
- * Upload a part in a multipart upload transaction
- * @param {String} name the object name
- * @param {String} uploadId the upload id
- * @param {Integer} partNo the part number
- * @param {File} file upload File, whole File
- * @param {Integer} start  part start bytes  e.g: 102400
- * @param {Integer} end  part end bytes  e.g: 204800
- * @param {Object} options
- */
-proto.uploadPart = function* uploadPart(name, uploadId, partNo, file, start, end, options) {
-  var data = {
-    stream: this._createStream(file, start, end),
-    size: end - start
-  };
-  return yield this._uploadPart(name, uploadId, partNo, data, options);
-};
-
-/**
- * Complete a multipart upload transaction
- * @param {String} name the object name
- * @param {String} uploadId the upload id
- * @param {Array} parts the uploaded parts
- *          - {Object} part the result from uploadPartCopy
- *            - {Integer} number partNo
- *            - {String} etag  part etag  uploadPartCopy result.res.header.etag
- * @param {Object} options
- */
-proto.completeMultipartUpload = function* completeMultipartUpload(name, uploadId, parts, options) {
-  parts.sort((a, b) => a.number - b.number);
-  var xml = '<?xml version="1.0" encoding="UTF-8"?>\n<CompleteMultipartUpload>\n';
-  for (var i = 0; i < parts.length; i++) {
-    var p = parts[i];
-    xml += '<Part>\n';
-    xml += '<PartNumber>' + p.number + '</PartNumber>\n';
-    xml += '<ETag>' + p.etag + '</ETag>\n';
-    xml += '</Part>\n';
-  }
-  xml += '</CompleteMultipartUpload>';
-
-  options = options || {};
-  options.subres = {uploadId: uploadId};
-  var params = this._objectRequestParams('POST', name, options);
-  params.mime = 'xml';
-  params.content = xml;
-  if (!(options.headers && options.headers['x-oss-callback'])) {
-    params.xmlResponse = true;
-  }
-  params.successStatuses = [200];
-  var result = yield this.request(params);
-
-  var ret = {
-    res: result.res,
-    bucket: params.bucket,
-    name: name,
-    etag: result.res.headers['etag']
-  };
-
-  if (options.headers && options.headers['x-oss-callback']) {
-    ret.data = JSON.parse(result.data.toString());
-  }
-
-  return ret;
-};
 
 is.file = function (file) {
   return typeof(File) !== 'undefined' && file instanceof File;
@@ -394,40 +245,6 @@ WebFileReadStream.prototype._read = function _read(size) {
   } else {
     this.readFileAndPush(size);
   }
-};
-
-/**
- * Upload a part in a multipart upload transaction
- * @param {String} name the object name
- * @param {String} uploadId the upload id
- * @param {Integer} partNo the part number
- * @param {Object} data the body data
- * @param {Object} options
- */
-proto._uploadPart = function* _uploadPart(name, uploadId, partNo, data, options) {
-  options = options || {};
-  options.headers = {
-    'Content-Length': data.size
-  };
-
-  options.subres = {
-    partNumber: partNo,
-    uploadId: uploadId
-  };
-  var params = this._objectRequestParams('PUT', name, options);
-  params.mime = options.mime;
-  params.stream = data.stream;
-  params.successStatuses = [200];
-
-  var result = yield this.request(params);
-
-  data.stream = null;
-  params.stream = null;
-  return {
-    name: name,
-    etag: result.res.headers.etag,
-    res: result.res
-  };
 };
 
 proto._createStream = function _createStream(file, start, end) {
