+++ lib/bucket.js
@@ -55,6 +55,11 @@ proto.useBucket = function useBucket(name, region) {
   return this;
 };
 
+proto.setBucket = function useBucket(name) {
+  this.options.bucket = name;
+  return this;
+};
+
 proto.getBucket = function getBucket() {
   return this.options.bucket;
 };
