+++ lib/client.js
@@ -103,10 +103,8 @@ merge(proto, require('./object'));
  * Bucket operations
  */
 merge(proto, require('./bucket'));
-/**
- * Multipart operations
- */
-merge(proto, require('./multipart'));
+//multipart upload
+merge(proto, require('./managed_upload'));
 /**
  * RTMP operations
  */
@@ -117,7 +115,10 @@ merge(proto, require('./rtmp'));
  */
 merge(proto, require('./common/multipart-copy.js'));
 merge(proto, require('./common/thunkpool.js'));
-
+/**
+ * Multipart operations
+ */
+merge(proto, require('./common/multipart'));
 /**
  * ImageClient class
  */
