+++ lib/common/multipart-copy.js
@@ -12,8 +12,8 @@ var proto = exports;
  * @param {Number} partNo the part number
  * @param {String} range  like 0-102400  part size need to copy
  * @param {Object} sourceData
- * @param {String} sourceData.sourceKey  the source object name
- * @param {String} sourceData.sourceBucketName  the source bucket name
+ *        {String} sourceData.sourceKey  the source object name
+ *        {String} sourceData.sourceBucketName  the source bucket name
  * @param {Object} options
  */
 proto.uploadPartCopy = function* uploadPartCopy(name, uploadId, partNo, range, sourceData, options) {
@@ -45,12 +45,12 @@ proto.uploadPartCopy = function* uploadPartCopy(name, uploadId, partNo, range, s
 /**
  * @param {String} name copy object name
  * @param {Object} sourceData
- * @param {String} sourceData.sourceKey  the source object name
- * @param {String} sourceData.sourceBucketName  the source bucket name
- * @param {Number} sourceData.startOffset  data copy start byte offset, e.g: 0
- * @param {Number} sourceData.endOffset  data copy end byte offset, e.g: 102400
+ *        {String} sourceData.sourceKey  the source object name
+ *        {String} sourceData.sourceBucketName  the source bucket name
+ *        {Number} sourceData.startOffset  data copy start byte offset, e.g: 0
+ *        {Number} sourceData.endOffset  data copy end byte offset, e.g: 102400
  * @param {Object} options
- * @param {Number} options.partSize
+ *        {Number} options.partSize
  */
 proto.multipartUploadCopy = function* multipartUploadCopy(name, sourceData, options) {
   this.resetCancelFlag();
@@ -209,8 +209,8 @@ proto._divideMultipartCopyParts = function _divideMultipartCopyParts(fileSize, p
  */
 proto._getObjectMeta = function* _getObjectMeta(bucket, name, options) {
   var currentBucket = this.getBucket();
-  this.useBucket(bucket);
+  this.setBucket(bucket);
   var data = yield this.head(name, options);
-  this.useBucket(currentBucket);
+  this.setBucket(currentBucket);
   return data;
 };
